SamplingUtil moved to plot.base.util package
