+++ plot-builder-portable/src/commonMain/kotlin/jetbrains/datalore/plot/builder/sampling/method/RandomStratifiedSampling.kt
@@ -8,7 +8,7 @@ package jetbrains.datalore.plot.builder.sampling.method
 import jetbrains.datalore.base.gcommon.base.Preconditions.checkArgument
 import jetbrains.datalore.plot.base.DataFrame
 import jetbrains.datalore.plot.builder.data.GroupUtil
-import jetbrains.datalore.plot.base.sampling.method.SamplingUtil
+import jetbrains.datalore.plot.base.util.SamplingUtil
 import jetbrains.datalore.plot.builder.sampling.GroupAwareSampling
 import kotlin.math.max
 import kotlin.math.round
